refactor: 피드백 반영 예외 처리 핸들러 구현
