+++ src/main/java/com/igemoney/igemoney_BE/attendance/controller/AttendanceController.java
@@ -4,12 +4,8 @@ import com.igemoney.igemoney_BE.attendance.dto.AttendanceResponseDto;
 import com.igemoney.igemoney_BE.attendance.service.AttendanceService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.NoSuchElementException;
-
 
 @RestController
 @RequiredArgsConstructor
@@ -21,14 +17,9 @@ public class AttendanceController {
 
     @GetMapping("/attendance")
     public ResponseEntity<?> attendance(@RequestHeader("Authorization") String authorizationHeader) {
-        try {
-            Long kakaoOauthId = extractOauthIdFromJwt(authorizationHeader);
-            AttendanceResponseDto response = attendanceService.getTodayAttendance(kakaoOauthId);
-            return ResponseEntity.ok(response);
-        } catch (NoSuchElementException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body("해당 유저를 찾을 수 없습니다.");
-        }
+        Long kakaoOauthId = extractOauthIdFromJwt(authorizationHeader);
+        AttendanceResponseDto response = attendanceService.getTodayAttendance(kakaoOauthId);
+        return ResponseEntity.ok(response);
     }
 
     private Long extractOauthIdFromJwt(String token) {
