+++ src/main/java/com/igemoney/igemoney_BE/attendance/exception/AttendanceNotFoundException.java
@@ -1,0 +1,7 @@
+package com.igemoney.igemoney_BE.attendance.exception;
+
+public class AttendanceNotFoundException extends Exception {
+    public AttendanceNotFoundException(String message) {
+        super(message);
+    }
+}
