+++ src/test/java/com/igemoney/igemoney_BE/attendance/controller/AttendanceControllerTest.java
@@ -1,0 +1,4 @@
+package com.igemoney.igemoney_BE.attendance.controller;
+
+public class AttendanceControllerTest {
+}
